+++ src/main/java/com/pwr/gastromate/dto/MenuItemIngredientDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -13,6 +15,6 @@ public class MenuItemIngredientDTO {
     private Integer menuItemId;
     private Integer ingredientId;
     private String ingredientName;  // Dodajemy nazwę składnika
-    private Float quantityRequired;
+    private BigDecimal quantityRequired;
     private Integer unitId;
 }
